--- dubbo-config/dubbo-config-api/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.config.ConfigPostProcessor
@@ -1,1 +1,0 @@
-feign-provider=org.apache.dubbo.config.processor.FeignClientAnnotationConfigPostProcessor
